Add an AnsDueDateView and Use it to refactor the AnsDueDateViewController
