--- Accompany/Resources/Accompany/Models/PopUpFormModel.swift
@@ -1,26 +1,0 @@
-//
-//  PopUpFormModel.swift
-//  Accompany
-//
-//  Created by Esperanza on 2023-05-12.
-//
-
-import Foundation
-import UIKit
-
-struct PopUpFormModel {
-  
-  // MARK: - Properties
-  
-  let questionTitle: String
-  
-  let questionDescriptionTitle: String
-  
-  let button: UIButton
-}
-
-
-struct DueDatePickerModel {
-  
-  let datePicker: UIDatePicker
-}
